Extract generic utilities from `utils' into `utils_io' and `utils_funcs' modules
`utils' should gradually gear towards an encapsulation of tensorflow
and DNN-related procedures; it should also be renamed accordingly in
the near future.
