+++ src/amplif.py
@@ -1,7 +1,8 @@
 # Basic amplification of input features
 
 from typing import *
-from utils import p1, tp1, np
+from utils_io import p1, tp1
+from utils import np
 from engine import Input, InputsDict, Bounds, TestTarget, Analyzer
 
 # ---
