+++ src/deepconcolic.py
@@ -7,6 +7,8 @@ sys.path.insert (0, os.path.join (__thisdir))
 
 import yaml
 from pathlib import Path
+from utils_io import *
+from utils_funcs import *
 from utils import *
 from bounds import UniformBounds, StatBasedInputBounds
 import fuzzer
