+++ src/filters.py
@@ -1,5 +1,6 @@
 # Input filters
 
+from utils_io import *
 from utils import *
 from engine import _InputsStatBasedInitializable, StaticFilter, Input
 from sklearn.neighbors import LocalOutlierFactor
