+++ src/pulp_norms.py
@@ -1,5 +1,5 @@
-from typing import *
 import numpy as np
+from typing import *
 from pulp import *
 from norms import LInf
 from lp import PulpLinearMetric
