+++ src/scripting.py
@@ -1,4 +1,4 @@
-from utils import *
+from utils_io import *
 import time
 
 def init_tics ():
