+++ src/utils.py
@@ -1,87 +1,20 @@
-from typing import *
 from abc import abstractmethod
-import os
+from utils_io import *
+from utils_funcs import *
 import sys
-import random
 import copy
-import datetime
-import numpy as np
-try:
-  import pandas as pd
-except:
-  pass
 
 # NB: importing cv2 and sklearn before tensorflow seems to solve an
 # issue with static TLS I've been having on an "oldish" version of
 # Linux (cf
 # https://github.com/scikit-learn/scikit-learn/issues/14485#issuecomment-633452991):
-import cv2
 import sklearn
 import tensorflow as tf
 from tensorflow import keras
 
-def rng_seed (seed: Optional[int]):
-  if seed is None:
-    seed = int (np.random.uniform (2**32-1))
-  print ('RNG seed:', seed) # Log seed to help some level of reproducibility
-  np.random.seed (seed)
-  # In case one also uses pythons' stdlib ?
-  random.seed (seed)
-
-def randint ():
-  return int (np.random.uniform (2**32-1))
-
 print ("Using TensorFlow version:", tf.__version__, file = sys.stderr)
 
-COLUMNS = os.getenv ('COLUMNS', default = '80')
-P1F = '{:<' + COLUMNS + '}'
-N1F = '\n{:<' + COLUMNS + '}'
-
-def tp1(x):
-  print (P1F.format(x), end = '\r', flush = True)
-
-def ctp1(x):
-  print (N1F.format(x), end = '\r', flush = True)
-
-def np1(x):
-  print (x, end = '', flush = True)
-
-def cnp1(x):
-  print ('\n', x, sep = '', end = '', flush = True)
-
-def p1(x, **k):
-  print (P1F.format(x), **k)
-
-def c1(x):
-  print (x)
-
-def cp1(x, **k):
-  print (N1F.format(x), **k)
-
-
-def xtuple(t):
-  return t if len(t) > 1 else t[0]
-
-def xlist(t):
-  return [t] if t is not None else []
-
-def seqx(t):
-  return [] if t is None else t if isinstance (t, (list, tuple)) else [t]
-
-def id(x):
-  return x
-
-def some(a, d):
-  return a if a is not None else d
-
-def appopt(f, a):
-  return f(a) if a is not None else a
-
-def s_(i):
-  return i, 's' if i > 1 else ''
-
-def is_are_(i):
-  return i, 'are' if i > 1 else 'is'
+# ---
 
 #the_dec_pos=0
 MIN=-100000
@@ -163,112 +96,6 @@ def activation_is_relu(layer):
 
 # ---
 
-def setup_output_dir (outs, log = True):
-  if not os.path.exists (outs):
-    os.makedirs (outs)
-  if not outs.endswith ('/'):
-    outs += '/'
-  if log: print ('Setting up output directory: {0}'.format (outs))
-  return outs
-
-# def setup_report_files (outs, ident, suff0 = '', suff = '.txt', log = True):
-#   if not os.path.exists(outs):
-#     sys.exit ('Output directory {0} was not initialized (internal bug)!'
-#               .format (outs))
-#   f = outs+ident+suff
-#   if log: print ('Reporting into: {0}'.format (f))
-#   return f, ident
-
-class OutputDir:
-  '''
-  Class to help ensure output directory is created before starting any
-  lengthy computations.
-  '''
-  def __init__(self, outs = '/tmp', log = None,
-               enable_stamp = True, stamp = None, prefix_stamp = False):
-    self.dirpath = setup_output_dir (outs, log = log)
-    self.enable_stamp = enable_stamp
-    self.prefix_stamp = prefix_stamp
-    self.reset_stamp (stamp = stamp)
-
-  def reset_stamp (self, stamp = None):
-    self.stamp = datetime.datetime.now ().strftime("%Y%m%d-%H%M%S") \
-                 if stamp is None and self.enable_stamp else \
-                 stamp if self.enable_stamp else ''
-
-  @property
-  def path(self) -> str:
-    return self.dirpath
-
-  def filepath(self, base, suff = '') -> str:
-    return self.dirpath + base + suff
-
-  def stamped_filename(self, base, sep = '-', suff = '') -> str:
-    return ((self.stamp + sep + base) if self.enable_stamp and self.prefix_stamp else \
-            (base + sep + self.stamp) if self.enable_stamp else \
-            (base)) + suff
-
-  def stamped_filepath(self, *args, **kwds) -> str:
-    return self.dirpath + self.stamped_filename (*args, **kwds)
-
-  def subdir(self, name) -> str:
-    dirname = self.filepath (name)
-    if not os.path.exists (dirname):
-      os.makedirs (dirname)
-    return dirname
-
-  def fresh_dir(self, basename, suff_fmt = '-{:x}', **kwds):
-    outdir = self.filepath (basename + suff_fmt.format (random.getrandbits (16)))
-    try:
-      os.makedirs (outdir)
-      return OutputDir (outdir, **kwds)
-    except FileExistsError:
-      return self.fresh_dir (basename, suff_fmt = suff_fmt, **kwds)
-
-# ---
-
-def _write_in_file (f, mode, *fmts):
-  f = open (f, mode)
-  for fmt in fmts: f.write (fmt)
-  f.close ()
-
-def write_in_file (f, *fmts):
-  _write_in_file (f, "w", *fmts)
-
-def append_in_file (f, *fmts):
-  _write_in_file (f, "a", *fmts)
-
-def save_in_csv (filename):
-  def save_an_array(arr, name, directory = './', log = True):
-    if not directory.endswith('/'): directory += '/'
-    f = directory + filename + '.csv'
-    if log: print ('Appending array into {0}'.format (f))
-    with open (f, 'a') as file:
-      file.write (name + ' ')
-      np.savetxt (file, arr, newline = ' ')
-      file.write ('\n')
-
-    # append_in_file (f, name, ' ', np.array_str (arr, max_line_width = np.inf), '\n')
-  return save_an_array
-
-def save_an_image(im, name, directory = './', log = True, channel_upscale = 255):
-  if not directory.endswith('/'): directory += '/'
-  f = directory + name + '.png'
-  if log: print ('Outputing image into {0}'.format (f))
-  cv2.imwrite (f, im * channel_upscale)
-
-def save_an_image_(channel_upscale = 255):
-  return lambda *args, **kwds: \
-    save_an_image (*args, channel_upscale = channel_upscale, **kwds)
-
-def save_adversarial_examples(adv, origin, diff, di):
-  save_an_image(adv[0], adv[1], di)
-  save_an_image(origin[0], origin[1], di)
-  if diff is not None:
-    save_an_image(diff[0], diff[1], di)
-
-# ---
-
 class cover_layert:
   pass
 
@@ -339,21 +166,6 @@ def get_cover_layers (dnn, constr, layer_indices = None,
 
 # ---
 
-def validate_strarg (valid, spec):
-  def aux (v, s):
-    if s is not None and s not in valid:
-      raise ValueError ('Unknown {} `{}\' for argument `{}\': expected one of '
-                        '{}'.format (spec, s, v, valid))
-  return aux
-
-def validate_inttuplearg (v, s):
-  if isinstance (s, tuple) and all (isinstance (se, int) for se in s):
-    return
-  raise ValueError ('Invalid value for argument `{}\': expected tuple of ints'
-                    .format (v))
-
-# ---
-
 # Do we really manipulate many DNNs at once?
 from functools import lru_cache
 @lru_cache(4)
@@ -403,11 +215,6 @@ def predictions (dnn, xl, top_classes = None):
 
 # ---
 
-def as_numpy (d):
-  if pd is not None:
-    return d.to_numpy () if isinstance (d, pd.core.frame.DataFrame) else d
-  else:
-    return d
 
 class raw_datat:
   def __init__(self, data, labels, name = 'unknown'):
@@ -605,169 +412,6 @@ def print_adversarial_distribution(advs, fname, int_flag=False):
                  .format(ave, std),
                  '#distance #accumulated adversarial examples fall into this distance\n',
                  *['{0} {1}\n'.format(xs[i], ys[i]) for i in range(0, len(xs))])
-
-
-# ---
-
-
-class Coverage:
-  """Basic helper class to manipulate and type-annotate coverage measures."""
-
-  def __init__(self, covered = None, total = None, non_covered = None):
-    if total != None:
-      self.total = total
-    elif covered != None and non_covered != None:
-      self.total = covered + non_covered
-    elif covered != None:
-      self.total = covered
-    elif non_covered != None:
-      self.total = non_covered
-    else:
-      self.total = 0
-
-    if covered != None:
-      self.c = covered
-    elif non_covered != None and self.total > 0:
-      self.c = self.total - non_covered
-    else:
-      self.c = 0
-
-
-  def __add__(self, x):
-    return Coverage (covered = self.c + x.c,
-                     total = self.total + x.total)
-
-
-  def __mul__(self, f: float):
-    return Coverage (covered = float(self.c) * f,
-                     total = self.total)
-
-
-  @property
-  def done(self) -> bool:
-    return self.total == self.c
-
-
-  @property
-  def as_prop(self) -> float:
-    return (((1.0 * self.c) / (1.0 * self.total))
-            if self.total != 0 else 0.0)
-
-
-  def __repr__(self):
-    return str(self.as_prop)
-
-
-# ---
-
-
-class Bounds:
-  """
-  Basic abstract class to represent any bounds.  (Mostly for typing
-  arguments and sub-classing.)
-  """
-
-  @property
-  def low (self) -> np.array(float):
-    raise NotImplementedError
-
-  @property
-  def up (self) -> np.array(float):
-    raise NotImplementedError
-
-  @abstractmethod
-  def __getitem__ (self, _idx: Tuple[int, ...]) -> Tuple[float, float]:
-    raise NotImplementedError
-
-
-# ---
-
-
-from collections import UserDict
-
-try:
-  # Use xxhash if available as it's probably more efficient
-  import xxhash
-  __h = xxhash.xxh64 ()
-  def np_hash (x):
-    __h.reset ()
-    __h.update (x)
-    return __h.digest ()
-except:
-  def np_hash (x):
-    return hash (x.tobytes ())
-  # NB: In case we experience too many collisions:
-  # import hashlib
-  # def np_hash (x):
-  #   return hashlib.md5 (x).digest ()
-
-class NPArrayDict (UserDict):
-  '''
-  Custom dictionary that accepts numpy arrays as keys.
-  '''
-
-  def __getitem__(self, x: np.ndarray):
-    return self.data[np_hash (x)]
-
-  def __delitem__(self, x: np.ndarray):
-    del self.data[np_hash (x)]
-
-  def __setitem__(self, x: np.ndarray, val):
-    x.flags.writeable = False
-    self.data[np_hash (x)] = val
-
-  def __contains__(self, x: np.ndarray):
-    return np_hash (x) in self.data
-
-
-# ---
-
-
-D, C = TypeVar ('D'), TypeVar ('C')
-
-
-class LazyLambda:
-  '''
-  Lazy eval on an unknown domain.
-  '''
-
-  def __init__(self, f: Callable[[D], C], **kwds):
-    super ().__init__(**kwds)
-    self.f = f
-
-  def __getitem__(self, x: D) -> C:
-    return self.f (x)
-
-  def __len__(self) -> int:
-    return self.f (None)
-
-
-class LazyLambdaDict (Dict[D, C]):
-  '''
-  Lazy function eval on a fixed domain.
-  '''
-
-  def __init__(self, f: Callable[[D], C], domain: Set[D], **kwds) -> Dict[D, C]:
-    super ().__init__(**kwds)
-    self.domain = domain
-    self.f = f
-
-  def __getitem__(self, x: D) -> D:
-    if x not in self.domain:
-      return KeyError
-    return self.f (x)
-
-  def __contains__(self, x: D) -> bool:
-    return x in self.domain
-
-  def __iter__(self) -> Iterator[D]:
-    return self.domain.__iter__ ()
-
-  def __setitem__(self,_):
-    raise RuntimeError ('Invalid item assignment on `LazyLambdaDict` object')
-
-  def __delitem__(self,_):
-    raise RuntimeError ('Invalid item deletion on `LazyLambdaDict` object')
 
 
 # ---
